refactor: command 패키지에 있는 command 파일들 dto 폴더 안으로 이동
