+++ src/main/java/com/umc/product/curriculum/adapter/in/web/v1/dto/request/SelectBestWorkbookRequest.java
@@ -1,6 +1,6 @@
 package com.umc.product.curriculum.adapter.in.web.v1.dto.request;
 
-import com.umc.product.curriculum.application.port.in.command.SelectBestWorkbookCommand;
+import com.umc.product.curriculum.application.port.in.command.dto.SelectBestWorkbookCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "베스트 워크북 선정 요청")
